3D refactoring and updated textures
