+++ backend/ShootTheMoon/ShootTheMoon/Properties/launchSettings.json
@@ -3,7 +3,7 @@
     "windowsAuthentication": false,
     "anonymousAuthentication": true,
     "iisExpress": {
-      "applicationUrl": "http://localhost:52243",
+      "applicationUrl": "http://localhost:8000",
       "sslPort": 44351
     }
   },
@@ -25,7 +25,7 @@
         "ASPNETCORE_ENVIRONMENT": "Development"
       },
       "dotnetRunMessages": "true",
-      "applicationUrl": "https://localhost:5001;http://localhost:5000"
+      "applicationUrl": "https://localhost:8005;http://localhost:8000"
     },
     "Docker": {
       "commandName": "Docker",
