+++ frontend/shoot-client/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@babylonjs/core": "4.2.0",
     "@babylonjs/gui": "4.2.0",
+    "@babylonjs/loaders": "^4.2.0",
     "@material-ui/core": "^4.11.3",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
